feat: qr 코드 생성 및 모달
