+++ packages/sdk/package.json
@@ -10,6 +10,7 @@
   },
   "devDependencies": {
     "@rollup/plugin-typescript": "^12.1.2",
+    "@types/qrcode": "^1.5.5",
     "@typescript-eslint/eslint-plugin": "^8.32.0",
     "@typescript-eslint/parser": "^8.32.0",
     "eslint": "^9.26.0",
@@ -17,5 +18,8 @@
     "rollup-plugin-dts": "^6.2.1",
     "typescript": "~5.8.3",
     "vite": "^6.3.5"
+  },
+  "dependencies": {
+    "qrcode": "^1.5.4"
   }
 }
