+++ packages/sdk/src/index.ts
@@ -1,11 +1,2 @@
-import { loader } from "./loader";
-import { requestPayment } from "./requestPayment";
-import type { PaymentReqBody } from "./requestPayment";
-
-const PaymentSDK = {
-  loader,
-  requestPayment,
-};
-
-export default PaymentSDK;
-export type { PaymentReqBody };
+export { loader } from "./loader";
+export { requestPayment } from "./requestPayment";
