+++ packages/sdk/src/loader/index.ts
@@ -8,15 +8,15 @@ export interface ValidationResponse {
 export async function loader({ key }: { key: string }): Promise<void> {
   if (!key) throw new Error("[PaymentSDK] key is required");
 
-  const res = await fetch("https://passion-pay-api.com/validate-sdk", {
+  const response = await fetch("https://passion-pay-api.com/validate-sdk", {
     method: "POST",
     headers: { "Content-Type": "application/json" },
     body: JSON.stringify({ key }),
   });
 
-  const json: ValidationResponse = await res.json();
+  const json: ValidationResponse = await response.json();
 
-  if (!res.ok || !json.valid) {
+  if (!response.ok || !json.valid) {
     throw new Error("[PaymentSDK] Invalid SDK key");
   }
