+++ packages/sdk/src/requestPayment/generateQR.ts
@@ -1,0 +1,6 @@
+import QRCode from "qrcode";
+
+export async function generateQR(token: string) {
+  const qrCodeUrl = await QRCode.toDataURL(token);
+  return qrCodeUrl;
+}
