+++ packages/sdk/src/requestPayment/type.ts
@@ -1,0 +1,10 @@
+export interface PaymentReqBody {
+  amount: number;
+  merchant_order_id: string;
+  merchant_id: string;
+}
+
+export interface PaymentResData {
+  transaction_token: string;
+  expires_at: string;
+}
